[plan-2026-07-31-b63a35aa/iter-1/step-5] widen the sincos embedding to every dtype policy, and fix the second hard-float32 cast that kept SupernodePooling dead
G-10 needed BOTH halves to be closed, so they ship together (D-002) with two
independent RED proofs:

  A. ContinuousSinCosEmbed.call forced `coords` to a literal "float32" and
     multiplied it against `omega`, which Keras autocasts to the compute dtype.
     Measured at HEAD: InvalidArgumentError [Op:Mul] at mixed_float16, float64
     AND mixed_bfloat16 -- the layer and all five of its consumer surfaces were
     dead at 3 of the 4 standard policies. Now computed in a never-narrow work
     dtype, with `omega` cast to it and the result cast back to compute_dtype.

  B. SupernodePooling._aggregate_messages normalised by a neighbour count it
     hard-cast to "float32" against a compute-dtype numerator. With A applied it
     still died at the same 3 policies with a DIFFERENT exception
     (TypeError: `x` and `y` must have the same dtype). The divisor is now built
     at the numerator's dtype, copying the file's own `_compute_messages` idiom.

I-B holds: the float32 output is bit-identical to HEAD on all 3 corpora,
including the `padding > 0` branch. Four dead-component injections were proven
RED for the right assertion (zeros 0.99999912, swapped sin/cos halves 1.4142,
the fix reverted = a raise at exactly 3 of 4, the divisor reverted = exactly the
6 SupernodePooling cells). Tolerances are per-policy, rtol=0, and measured at a
stated coord regime; fp16/bf16 large-position accuracy is NOT claimed fixed.

Gates (CPU, serial): embedding+supernode 303p/2xf (284 baseline +19 new cases),
geometric 241p/2xf (233 +8), text encoder+decoder 183p (unchanged).

+++ src/dl_techniques/layers/geometric/supernode_pooling.py
@@ -397,9 +397,21 @@ class SupernodePooling(keras.layers.Layer):
         # Sum messages for each supernode
         aggregated = ops.sum(messages, axis=0)  # (num_supernodes, hidden_dim)
 
-        # Normalize by number of neighbors to get mean
-        neighbor_counts = ops.sum(ops.cast(neighbor_mask, "float32"), axis=0, keepdims=True)  # (1, num_supernodes)
-        neighbor_counts = ops.maximum(neighbor_counts, 1.0)  # Avoid division by zero
+        # Normalize by number of neighbors to get mean.
+        # DECISION plan-2026-07-31T210633-b63a35aa/D-002
+        # The divisor is built at the NUMERATOR's dtype, copying this file's own
+        # sibling idiom `ops.cast(neighbor_mask, messages.dtype)` in
+        # `_compute_messages`. Do NOT restore the hard `"float32"` literal (nor the
+        # bare `1.0` Python float, which `ops.maximum` also resolves to float32): the
+        # numerator is at the compute dtype, so the divide raised
+        # `TypeError: 'x' and 'y' must have the same dtype, got tf.float16 != tf.float32`
+        # under mixed_float16, float64 and mixed_bfloat16. This is the SECOND of the
+        # two hard dtype literals on the sincos path; fixing only
+        # `ContinuousSinCosEmbed` left this layer dead at 3 of the 4 policies.
+        neighbor_counts = ops.sum(
+            ops.cast(neighbor_mask, aggregated.dtype), axis=0, keepdims=True)  # (1, num_supernodes)
+        neighbor_counts = ops.maximum(
+            neighbor_counts, ops.cast(1.0, aggregated.dtype))  # Avoid division by zero
 
         aggregated = aggregated / ops.transpose(neighbor_counts)  # (num_supernodes, hidden_dim)
